requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
youtube-transcript-api
googletrans==4.0.0rc1
PyMuPDF
python-docx
